refactor with jen, deleted reviewmodel
