+++ CoffeeReview/Models/Coffee.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace CoffeeReview
+namespace CoffeeReview.Models
 {
     public class Coffee
     {
