+++ CoffeeReview/Repositories/IRepository.cs
@@ -1,6 +1,10 @@
-﻿namespace CoffeeReview.Repositories
+﻿using System.Collections.Generic;
+
+namespace CoffeeReview.Repositories
 {
-    public interface IRepository<T>
+    public interface IRepository<T> where T : class
     {
+        IEnumerable<T> GetAll();
+        T GetById(int id);
     }
 }
