+++ CoffeeReview/wwwroot/css/style.css
@@ -1,0 +1,2 @@
+﻿body {
+}
